+++ changelog.md
@@ -1,3 +1,5 @@
 #initial update: v0.0.1
 
 - basic setup
+- axios ongoing
+- api playground ongoing